requirements.txt
aiofiles==23.2.1
aiogram==3.18.0
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.4.4
pillow==11.2.1
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.0
yarl==1.20.1
